Simplify binary operations
Previously b'1' and b'0', i.e. bytes and not bits, were used to
construct the compressed output. Conversions from this list of
bytes representing strings, needed to be done.

Started now using the bitarray data structure to construct the binary
output in a more efficient, cleaner and simpler way.

The compressed output is still the same, the code passes the tests where
the compressed text is checked against a pre-computed compressed output.