Rename main.py -> cli.py, add new main.py
Renamed the `main.py` located within the `compressor` package into
`cli.py`, as well as the function `main()` into `run()`. Added a new
proper main.py outside the package which imports and runs the cli.
Also updated the poetry script for running the compressor to reflect
these changes.

The main.py file was not supposed to be actually run directly before,
as relative imports would not work. The poetry script, which was
previosly the only way to run the projcet, but now it can also be ran
with `python main.py` (within `poetry shell` to have deps available).