Improve error handling
Added a decorator for the `run()` function in cli.py` which handles
errors from the `Compressor`. It handles `CompressionError`s which are
explicitly thrown by `Compressor`, but also reports any unexpected
non-`CompressionError` exceptions.

Added a decorator for the compression and decompression methods in
`Compressor`, that also does error handling but also checks for existing
output files.