Fix bug in Huffman and refactor
Moved methods more closely related to handling the Huffman tree, from
Huffman into HuffmanTreeNode.

Fixed a bug in the Huffman implementation which happened when
compressing text contents that only contained a single char. Bug was in
Huffman coding, where if there was only a single character, i.e. the
Huffman tree consisted of only a single node (that is root and leaf at
the same time), it would not have been allocated a code -> compressed
output was very short and invalid.